telegram_listener/requirements.txt
telethon==1.32.0
python-dotenv==1.0.1
websockets==12.0
asyncio==3.4.3
WEBSOCKET_URL = https://websocket-server-vubd.onrender.com
